refactor cursor/buffer display copy
